+++ src/App.js
@@ -9,7 +9,7 @@ import { ActiveStepProvider, PinsProvider, SessionProvider } from './storage/con
 import Landing from './ui/pages/Landing';
 import Content from './ui/pages/Content';
 import Completion from './ui/pages/Completion';
-import Review from './ui/pages/Review/Landing';
+import Review from './ui/pages/Review';
 import Home from './ui/pages/Home';
 //tbr import CORsTestButtons from './ui/layout/CORsTestButtons';
 import { store } from "./storage/store";
