+++ src/ui/pages/Completion.jsx
@@ -57,6 +57,7 @@ const useStyles = makeStyles((theme) => ({
     }
 }));
 
+// completion page
 const Completion = () => {
     const classes = useStyles();
 	return (
